MP4 reader: always process whole key frame table
The old code used a way of iterating over the table that would get hung
up on duplicate entries. Additionally it relied on the table being
sorted in ascending order in the source file.

The result in such a case was that only key frames up to and including
the first duplicate frame index in the key frame index table were marked
as key frames in the output file, even though more key frames index
entries might exist.

The new method simply iterates over the whole table once, from start to
finish, and looks the referenced frames up properly.

This is the last part of the fix for #1867.