APE Tags: fix size parsing and add support for APEv1 tags
What we're reading is the footer, mandatory in all APEv1 files and in all
APEv2 files where the tags are at the end.
The header however is nonexistant in APEv1 and optional in APEv2 if the
tags are at the end of the file, so check its presence by looking at the
APE Tag flags instead of assuming it's always there.

Remove a now unnecessary function while at it as no APEv1 specific function
will be needed.

Signed-off-by: James Almer <jamrial@gmail.com>