mkvmerge: account enqueued bytes even for deferred packets
When appending a file, the packets for packetizers who haven't
finished their old file yet are put into a deferred queue. This
happens with readers that don't offer access to arbitrary tracks and
are read from start to finish resulting in packets for packetizers
which weren't pulled.

Before the patch those deferred packets weren't counted for the number
of enqueued bytes. The reader could therefore not judge properly
whether or not to hold a packetizer when a read request comes in.

This is a prerequisite for fixing #1893. As it affects non-Matroska
files, too, it deserves its own commit.