HEVC parser: only use last provided timestamp for a given stream position
In certain cases there are several timestamps from the source
container queued up for a given position of the NALU in the stream. In
such cases using the first available timestamp will result in
audio/video desync. Instead the last timestamp whose stream position
is smaller than or equal to the NALU's stream position should be used.

Fixes the HEVC equivalent of the problem with AVC described in #1908.