AVC/h.264 (framed) output: honor forced default duration signaling interlaced content
Internally the default duration given on the command line is stored as
the duration of a progressive frames. Additionally the framed
AVC/h.264 output module doesn't actually check whether or not the
current block contains a frame or a field. This combination leads to
the situation that specifying a default duration that signals
interlacing (e.g. 50i) results in an actual default duration of 40ms,
that of a progressive frame.

This change passes the information provided by the user about frame
vs. fields from the command line through to the output module so that
it can react accordingly.

Fixes #1916.