HEVC parser: fix calculation of number of parameter set arrays in HEVCC
The number of parameter set arrays is not the sum of the number of
VPS, SPS, PPS and SEI NALUs, but the number of different types. For
example, if there's one VPS, one SPS, two PPS and no SEI NALUs, the
number of parameter set arrays must be three and not four.

Fixes the video-related part of #1938.