HEVC parser: don't parse bitstream restriction twice if timing info is present in VUI
In order to fix #1924, I added bitstream restriction handling code in
the VUI parser in commit 2a385ab1ecc12c81b5977cb6674945284fcfe867.
Unfortunately I didn't realize that the code was present but in the
wrong place. It was only called if timing information was present,
too.

The result of commit 2a385ab1ecc12c81b5977cb6674945284fcfe867 was that
the bitstream restriction was now handled twice if timing information
was present.

The superfluous and wrongfully-placed copy has been removed. This
fixes #1924 properly. It also fixes #1958.