MPEG TS: don't emit charset warning for ISO 6937
The iconv version on macOS doesn't support that encoding. At the
moment mkvmerge only requires that encoding when reading the station
names from MPEG transport streams, and those are only shown to the
user as a help for deciding which tracks to select. Therefore the
information isn't critical, and failure to decode it properly doesn't
warrant a warning.

Fixes #2023.