GUI: iterate over all QStandardItems after drag & drop
It seems that dragging & dropping sometimes leaves Qt's internal state
somewhat foobared if at least one of the columns is hidden at the time
the items are dropped. This causes subsequent drag attempts to
segfault in the "start drag" function trying to serialize the standard
items' states.

For some reason iterating over all rows, all columns, for all parents
in the model and requesting the corresponding QStandardItem fixes the
internal state to the point that dragging doesn't crash anymore.

Fixes #2009.