GUI: chapters: add "calculate & set end timestamps" functionality
The function is available from the "additional modifications" dialog.

For most entries the smallest start timestamp of all chapters on the
same level higher than the current chapter's start timestamp will be
used as its end timestamp.  If there is no such chapter, the parent
chapter's end timestamp will be used instead.

If the chapters were loaded from a Matroska file, the end timestamp
for very last chapter on the top-most level will be derived from the
file's duration.

Implements #1887.