MKVToolNix GUI: removed keyboard shortcuts for switching between tools
This removes shortcuts such as `Ctrl+Alt+1` for switching to the
multiplexer. These shortcuts overlapped with basic functionality on
keyboards that use an `AltGr` key, e.g. German ones, where `AltGr+7`
emits `{`. As `AltGr+key` is implemented as `Ctrl+Alt+key` under the
hood, this means that `AltGr+7` is really `Ctrl+Alt+7` which the GUI
now took to mean "switch to the job queue" instead of "insert
`{`".

Fixes #2056.