GUI: mux: menu entry for setting destination file name from source file's
This is a new menu entry in the "source files" list box's context
menu.  Selecting it will force the GUI to consider the selected file
to be the reference for automatically setting the file name, no matter
which file was originally added as the first file.

It will also force setting the destination file name once if automatic
destination file name generation is turned off in the preferences.

Implements part of #2058.