GUI: headers: refactor page initialization to use `property_element_c`
The information about each element's parents, its type, title and
description is available in property_element_c and used by mkvpropedit
already. As that's all the information needed by the GUI's header
editor, there's no reason to duplicate it in the header
editor. Instead it can iterate over the tables provided by
`property_element_c` and create pages based on that information.

This is in preparation of supporting the changes required for #2038
and #2064.