propedit: store (sub-)sub-sub-masters in change, not track
One track can only have one type of sub-master that changes may
affect: TrackVideo or TrackAudio, and there are no other masters on
that level that are relevant. So the sub-master can still be stored in
the track.

Below that there are now at least two different masters changes for
the same track may apply to: VideoColour and VideoProject. Therefore
storing the sub-sub-masters and its children in the track doesn't make
sense.

This is in preparation of adding support for "video projection" track
header elements in #2064.