all: unset LANGUAGE env var if user requests translation
GNU gettext treats the `LANGUAGE` environment variable as the
authoritative source for which languages can be selected. If it is
set, then setting `LC_MESSAGES` to a language not listed in `LANGUAGE`
doesn't have an effect — GNU gettext will then select the first
existing from `LANGUAGE`.

However, command-line options and (in the case of the GUI)
configuration files should take precedence over the environment
variables. Therefore the translation routines in all MKVToolNix
programs have to un-set `LANGUAGE` if the user explicitly tells the
programs to use a certain language.

Fixes #2070.