AAC reader: show error if sampling frequency or channels is 0
The problem is with the ADTS AAC container format itself. The ADTS
container cannot contain all the information modern AAC
requires. Therefore the LOAS/LATM container was invented.

For example: in ADTS there are only two bits for the profile which
doesn't allow for signalling HE-AACv2. Another problem is the number
of channels for which only three bits are available.

For files that have all those channel bits set to 0, the output file
contained a channel count of 0 — obviously invalid.

See #2107.