GUI: headers: don't compare original/current value if element's not present
This mostly fails on timestamp elements as the original value is an
invalid date and the current value is the minimum that Qt's
`QDateTime` control allows at that moment. That's usually something
like 2000-01-01 00:00:00.

However, if the element isn't present in the file, then it doesn't
matter what the current value in the GUI is — it won't be written
anyway. Therefore it shouldn't be used in the decision whether or not
to ask the user to confirm discarding unsaved changes.

Fixes #2167.