rename README.Windows.md to README.md
The file is solely about compiling for Windows, not for using it on
Windows or anything else. SO make it clearer.

A second advantage is that GitLab won't get confused. GitLab
automatically places a link called "ReadMe" on the overview page. It
links to the first file matching the glob "README*" in alphabetical
order. As "README.Windows.md" sorts lower than "README.md", the link
was pointing to a file explaining to the user how to compile for
Windows. Not the ideal landing page for a repository.