mm_read_buffer_io_c: retrieve file position on object creation
When an `mm_read_buffer_io_c` instance was created, it assumed the
underlying I/O object's stream position was 0. This doesn't have to be
the case, though. As the read buffer I/O class caches the file
position, too, this lead to it returning data from the wrong portion
of the underlying I/O object on subsequent reads.

One observable issue was that trying to detect the file type of MPLS
files that refer to short M2TS files failed with segmentation
faults.