GUI: mux: disable tracks by default when appending to disabled tracks
For example: you add a file, disable the second audio track. Then you
append another file (again having two audio tracks). The second audio
track from the appended file will end up appended to the second audio
track from the first file by default — and as that is disabled, the
appended second audio track should be disabled too.

This is even clearer when considering that the GUI doesn't allow the
user to enable an appended track that's appended to a disabled track.