AC-3: change dialog normalization removal to set value to 31
In AC-3 the `dialnorm` parameter states at which level the audio was
encoded it: as -dB from 100%. A higher value (e.g. 10 instead of 5)
means that the audio must be amplified more in order to bring it back
to 100%. Therefore setting the value to 1 means -1 dB from 100% which
in turn means that it'll be amplified very little on playback and
remain pretty quiet.

eac3to sets the value to 31 instead (meaning -31 dB from 100%, the
player has to amplify a lot, the result will be a lot louder). So
let's match what it does.