mkvmerge: set 'doc type (read) version' depending on elements written
Prior to this change both fields were often set according to the
features that were enabled. For example, the 'doc type read version'
was set to 2 no matter whether or not a `SimpleBlock` element was
actually used unless the `--engage no_simpleblocks` option was used.

Now the decision is based solely on which elements and which values
are written.

Part of the fix of #2240.