FLAC: skip/ignore ID3 tags
The FLAC specs specifically state that the only tag/comment system is
the FLAC comment header. It also states that the reference FLAC
utility can skip ID3 tags in FLAC files — but only as a courtesy, that
they're not part of the spec and that other utilities will likely not
work with such files.

mkvmerge now works with such files, too.

Implements #2243.