GUI: info: implement scanning and reading clusters
All elements starting with and after the first cluster are only loaded
partially (only their element head). Their child elements are loaded
on demand: when the user expands that element in the GUI. On
collapsing the element the children are deleted again in order to keep
memory usage down.

Implements the rest of #2104.