Preserve Video Colour Space value for uncompressed video tracks
MKV files generated with GStreamer matroskamux for uncompressed video contain value in ColourSpace element.
If such file is edited with mkvmerge or mkvtoolnix-gui resulting file can't be decoded anymore by GStreamer.
Root cause is that GStreamer relies on presence of this element to chose pixel format for uncompressed video.

According to https://github.com/Matroska-Org/matroska-specification/blob/master/ebml_matroska.xml
'This Element is MANDATORY in TrackEntry when the CodecID Element of the TrackEntry is set to "V_UNCOMPRESSED"'

Signed-off-by: Andrey Vostrikov <andrey.vostrikov@cogentembedded.com>