mkvextract: verify first element found is an EBML Head
`FindNextID` considers anything that looks like an ID to be valid. If
the resulting ID is neither a known global nor an element from the
passed context, a Dummy Element is returned instead.

Therefore mkvextract must verify that the element found is indeed an
EBM Head element in order to determine whether or not the source file
is actually a Matroska/WebM file.

Fixes #2281.