bit_writer_c: change put_bit() back to taking a bool
Taking an `int` was done due to clang-tidy's modernize checks
bemoaning the use of integers 0 and 1 instead of `true` and `false`
for Boolean parameters. However, that had the unintended consequence
of truncating values if the values given to the parameter is e.g. a
`uint64_t` for which a bit was tested with "value & (1 << shift)" —
which relied on implicit conversion to bool.