Matroska reader: don't downshift timestamps twice when appending
Back in the day the Matroska reader was only trying to shift
timestamps to 0 by using the first cluster's cluster timestamp. Quite
a while later real zeroing was implemented by actually reading the
timestamps of all blocks of the first ~20 MB of the file or so, but
the original method wasn't removed.

This was bad in two ways:

1. The older method was wrong as the first cluster's timestamp didn't
   have to be equal to the lowest timestamp in the file as block
   timestamps can be negative.
2. There were two methods for downshifting that were applied on top of
   each other when appending.

This removes the old, wrong way.

Fixes #2345.