GUI: info: fix position of "cues" in tree
The code handling cues behaves differently whether it's run as part of
MKVToolNix GUI or the standalone mkvinfo. In the latter case, its
sub-entries will only be shown in verbosity level 2 or
higher. However, in GUI mode they must always be shown.

Additionally, this caused the cues to be shown at a wrong location:
inside the previous master element instead of on level 1.

Fixes #2361.