GUI: mux: set chapter name template from preferences; pass to mkvmerge if needed
There are now two situations in which mkvmerge generates names for
chapters:

1. If generation is enabled (either in intervals or for each appended
   file)
2. When reading chapters from MPLS playlists

The GUI will now pass the chapter name template to mkvmerge in both
situations, even if the template is empty (which will result in
chapters being generated without names).

In order to provide a sensible default, the chapter name template
property will be set to the value in the preferences (section "Chapter
editor").

Part of the implementation of #2275.