GUI: cache default font
The function determining the default user interface font was leaking
font resources on Windows. This in itself isn't a huge
problem. However, the function determining the sidebar's font size is
called rather often, and that one calls the "default UI font"
function.

The result was memory exhaustion regarding font resources, a general
slowdown not only of MKVToolNix GUI but also other Windows
applications and subsequently a crash of MKVToolNix GUI.

Fixes #2372.