teletext-to-SRT: flush queued frame sooner if future merge is impossible
So far the "erase" flag wasn't used at all when determining whether or
not to flush the currently queued frame. It can be used, though, as
it's clear that erasing means the following content cannot be merged
to the current one.

Flushing earlier also means that fewer data will have to be buffered
before the frame is handed over to the output routines which will then
interleave frames from different tracks.

This also reverts the increase in maximum buffered amount of data in
the MPEG transport stream reader; it's back to 20 MB now.

Fixes #2393.