GUI: chapters: retain analyzer when opening Matroska file without chapters
The analyzer must be kept as it is needed when writing the chapters
back. This had always been the case until the implementation of #2439
where it got lost accidentally.

The result of not having the analyzer when saving is that the file is
truncated to a couple of KB, effectively destroying all of its content.

Fixes #2476.