GUI: use virtual destructors to job private classes
The pimpl idiom (data members stored in a private class) requires that
the main base class stores a pointer to the private class
instance. This pointer's type must be the private base
class. Obviously the pointer may point to an instance of a derived
private class.

Deletion happens via the private base class. If the destructor of that
class hierarchy isn't virtual, only the private base class's
destructor is called, even if the pointer points to a derived private
class. Consequently the destructors of data member in derived classes
aren't called either.

This is a problem for data members that need to clean up after
themselves, such as temporary files that should be deleted whenever
the

Fixes #2499.